Fix: Prevent post-processing complications from leading to a crash
Under more complicated scenarios Tcpflow sometimes has hick-ups which (might) lead to parsing complications in post-processing step. With this fix this shouldn't lead to a crash anymore. The problem content is still concated to the according flow and will as such be included in the output just as usual. Using "-v" (verbose) output one can take a look at the problem content for further analysis/investigation.